refactor(buffer): extract buffer-codec.ts (817 → 701 + 140)
Same per-concern split pattern. The `Buffer` class itself is one
cohesive `Uint8Array` subclass implementation, so the win comes
from extracting the two encoding-codec functions that aren't
class methods.

Layout:

  src/buffer.ts        (701 LoC) — Buffer + SlowBuffer classes +
                                    constants (was 817)
  src/buffer-codec.ts  (140 LoC) — encodeString(str, encoding) +
                                    decodeString(buf, encoding,
                                    start?, end?) — per-encoding
                                    switch logic for utf8/ascii/
                                    latin1/base64/base64url/hex/
                                    utf16le — plus checkOffset
                                    range-check helper

textEncoder / textDecoder module-locals now live in buffer-codec.ts.
The 2 remaining call sites in `Buffer.byteLength` use a one-shot
`new TextEncoder()` — TextEncoder is cheap to construct, and
byteLength is rarely on a hot path.

Public surface preserved verbatim; no exports changed.

326 GJS buffer tests still green; tsc + build clean.

STATUS.md: new Completed entry.
