+++ STATUS.md
@@ -405,6 +405,8 @@ Not yet implemented (but potentially relevant for GJS projects):
 
 ### Completed
 
+- ~~**`@gjsify/buffer` — extract `buffer-codec.ts` from `buffer.ts` (817 → 701 + 140)**~~✓ — Same per-concern split pattern (each module owns a clear slice). The Buffer class proper is one cohesive `Uint8Array` subclass implementation, so the win comes from extracting the two encoding-codec functions that aren't class methods. Layout: `buffer-codec.ts` (140 LoC) holds `encodeString(str, encoding)` and `decodeString(buf, encoding, start?, end?)` — the per-encoding String↔Uint8Array switch logic for utf8/ascii/latin1/base64/base64url/hex/utf16le — plus the small `checkOffset(offset, ext, length)` range-check helper. `buffer.ts` shrinks from 817 LoC to 701 LoC; its imports section is noticeably less cluttered (the `textEncoder` / `textDecoder` module-locals live in `buffer-codec.ts` now; the 2 remaining call sites in `Buffer.byteLength` use a one-shot `new TextEncoder()` — TextEncoder is cheap to construct, and `byteLength` is rarely on a hot path). Public surface preserved verbatim; no exports changed. 326 GJS buffer tests still green; tsc + build clean.
+
 - ~~**`README.md` modernized — `npm`/`yarn` commands replaced with `gjsify <subcommand>`**~~✓ — Top-level user-facing README brought in line with v0.4.21's Node-free CLI. Quick Start now leads with the `curl … install.mjs | gjs -m -` Node-free bootstrap; the legacy `npm install -g @gjsify/cli` is demoted to a clearly-marked escape hatch. New-project flow is `gjsify create` + `gjsify install --immutable` + `gjsify build` + `gjsify run start` (was `npm create @gjsify/app` + `npm install` + `npm run build` + `npm start`). Prerequisites section drops `nodejs` from the Fedora/Ubuntu apt/dnf one-liners and labels Node.js itself as optional (only needed for the cross-validation `test:node` track or for managing `@gjsify/cli` via `npm install -g`). Development section: `yarn build` → `gjsify foreach -A -t build`, `yarn check` → `gjsify check`, `yarn test` → `gjsify foreach -A test`, `yarn test:integration*` → per-suite `gjsify workspace <name> test`, per-package `yarn test:{node,gjs}` → `gjsify test [--runtime gjs|node]` with a paragraph explaining the new `gjsify test` runner. Verified the cited commands by running each of `gjsify --help` / `install --help` / `dlx --help` / `run --help` / `foreach --help` / `workspace --help` / `check --help` / `test --help` / `build --help` / `create --help` against the committed bundle.
 
 - ~~**`inline-static-reads` — refuse foreign-namespace `.join` evaluation + bail on `readdirSync({withFileTypes:true})`**~~✓ — Root-cause fix for a build-time inliner regression that crashed ts-for-gir's GJS bundle when running `doc`/`json` subcommands. `identifierName()` extracted the property name from any `MemberExpression` callee, so `dir.join('/')` (Array.prototype.join — typedoc's `discoverFiles` walks a string-array stack) was misread as `path.join('/')`. The static evaluator then resolved that to `'/'`, called `readdirSync('/', {withFileTypes:true})` at build time, and inlined the **build machine's root filesystem listing** (`['afs','bin','boot',...]`) as a string array. At runtime the consumer iterates that array and calls `child.isFile()`, which throws `i.isFile is not a function` because each element is a string, not a `Dirent`. Two layered fixes: (a) `identifierName()` now requires the MemberExpression object to be a known module namespace identifier (`path` or `fs`), so unrelated `.join()` / `.dirname()` / `.resolve()` etc. method calls bail out instead of resolving. (b) `readdirSync` inlining bails when the call passes `{withFileTypes:true}` — even with a correctly-evaluated path, we'd substitute `string[]` for the expected `Dirent[]`, breaking any consumer that uses `.isFile()`/`.isDirectory()`. Defence in depth against new static path-resolution heuristics we might add later. New `packages/infra/cli/src/inline-static-reads.spec.ts` locks the regression: 4 cases covering the typedoc shape, the `withFileTypes:true` bail-out, the legitimate `path.join`-resolved happy path that still inlines, and a generic `arr.join('/')` non-match. The bug only manifested in bundles whose source happened to use Array.prototype.join with a single string argument (typedoc 0.28's `discoverFiles`, possibly others) and whose static path arg coincidentally resolved to a real directory at build time — silent for everything else, which is how it stayed in master for the full @gjsify/cli 0.4.x train.
