+++ packages/node/buffer/src/buffer.ts
@@ -1,143 +1,27 @@
 // Native Buffer implementation for GJS
-// Reference: Node.js lib/buffer.js, Deno ext/node/polyfills/buffer.ts
-
-// BufferEncoding is a global type provided by @types/node
+// Reference: Node.js lib/buffer.js, Deno ext/node/polyfills/buffer.ts.
+//
+// Composition layout (see each module's header for details):
+//   - encoding.ts      — normalizeEncoding / checkEncoding / Encoding (moved
+//                        from `@gjsify/utils` in PR #265 — Phase 1 of the
+//                        utils audit).
+//   - base64.ts        — base64Encode / base64Decode / atobPolyfill /
+//                        btoaPolyfill (moved from `@gjsify/utils` same PR).
+//   - buffer-codec.ts  — encodeString / decodeString / checkOffset
+//                        (per-encoding String ↔ Uint8Array switch logic,
+//                        ~135 LoC). No Buffer-class coupling.
+//   - buffer.ts        — this file: Buffer + SlowBuffer classes + the
+//                        constants block (kMaxLength / kStringMaxLength /
+//                        constants).
+//
+// BufferEncoding is a global type provided by @types/node.
 
 import { normalizeEncoding, checkEncoding } from './encoding.js';
-import { atobPolyfill as _atob, btoaPolyfill as _btoa, base64Decode } from './base64.js';
+import { atobPolyfill as _atob } from './base64.js';
+import { encodeString, decodeString, checkOffset } from './buffer-codec.js';
 
-const textEncoder = new TextEncoder();
-const textDecoder = new TextDecoder();
-
-// SharedArrayBuffer may not be available in GJS
+// SharedArrayBuffer may not be available in GJS.
 const hasSharedArrayBuffer = typeof SharedArrayBuffer !== 'undefined';
-
-// normalizeEncoding and checkEncoding are imported from @gjsify/utils
-
-// Encode string to Uint8Array
-function encodeString(str: string, encoding: BufferEncoding): Uint8Array {
-  switch (encoding) {
-    case 'utf8':
-      return textEncoder.encode(str);
-
-    case 'ascii': {
-      const buf = new Uint8Array(str.length);
-      for (let i = 0; i < str.length; i++) {
-        buf[i] = str.charCodeAt(i) & 0x7f;
-      }
-      return buf;
-    }
-
-    case 'latin1': {
-      const buf = new Uint8Array(str.length);
-      for (let i = 0; i < str.length; i++) {
-        buf[i] = str.charCodeAt(i) & 0xff;
-      }
-      return buf;
-    }
-
-    case 'base64': {
-      const standard = str.replace(/-/g, '+').replace(/_/g, '/');
-      return base64Decode(standard);
-    }
-
-    case 'base64url': {
-      const base64 = str.replace(/-/g, '+').replace(/_/g, '/');
-      return encodeString(base64, 'base64');
-    }
-
-    case 'hex': {
-      const bytes = str.length >>> 1;
-      const buf = new Uint8Array(bytes);
-      for (let i = 0; i < bytes; i++) {
-        const hi = parseInt(str[i * 2], 16);
-        const lo = parseInt(str[i * 2 + 1], 16);
-        if (Number.isNaN(hi) || Number.isNaN(lo)) break;
-        buf[i] = (hi << 4) | lo;
-      }
-      return buf;
-    }
-
-    case 'utf16le': {
-      const buf = new Uint8Array(str.length * 2);
-      for (let i = 0; i < str.length; i++) {
-        const code = str.charCodeAt(i);
-        buf[i * 2] = code & 0xff;
-        buf[i * 2 + 1] = (code >> 8) & 0xff;
-      }
-      return buf;
-    }
-
-    default:
-      return textEncoder.encode(str);
-  }
-}
-
-// Decode Uint8Array to string
-function decodeString(buf: Uint8Array, encoding: BufferEncoding, start?: number, end?: number): string {
-  const slice = start !== undefined || end !== undefined
-    ? buf.subarray(start ?? 0, end ?? buf.length)
-    : buf;
-
-  switch (encoding) {
-    case 'utf8':
-      return textDecoder.decode(slice);
-
-    case 'ascii': {
-      let result = '';
-      for (let i = 0; i < slice.length; i++) {
-        result += String.fromCharCode(slice[i] & 0x7f);
-      }
-      return result;
-    }
-
-    case 'latin1': {
-      let result = '';
-      for (let i = 0; i < slice.length; i++) {
-        result += String.fromCharCode(slice[i]);
-      }
-      return result;
-    }
-
-    case 'base64': {
-      let binary = '';
-      for (let i = 0; i < slice.length; i++) {
-        binary += String.fromCharCode(slice[i]);
-      }
-      return _btoa(binary);
-    }
-
-    case 'base64url': {
-      const base64 = decodeString(slice, 'base64');
-      return base64.replace(/\+/g, '-').replace(/\//g, '_').replace(/=+$/, '');
-    }
-
-    case 'hex': {
-      let result = '';
-      for (let i = 0; i < slice.length; i++) {
-        result += slice[i].toString(16).padStart(2, '0');
-      }
-      return result;
-    }
-
-    case 'utf16le': {
-      let result = '';
-      for (let i = 0; i + 1 < slice.length; i += 2) {
-        result += String.fromCharCode(slice[i] | (slice[i + 1] << 8));
-      }
-      return result;
-    }
-
-    default:
-      return textDecoder.decode(slice);
-  }
-}
-
-function checkOffset(offset: number, ext: number, length: number): void {
-  if (offset + ext > length) {
-    throw new RangeError('Attempt to access memory outside buffer bounds');
-  }
-}
 
 const INSPECT_MAX_BYTES = 50;
 
@@ -288,7 +172,7 @@ export class Buffer extends Uint8Array {
     const enc = normalizeEncoding(encoding);
     switch (enc) {
       case 'utf8':
-        return textEncoder.encode(string).byteLength;
+        return new TextEncoder().encode(string).byteLength;
       case 'ascii':
       case 'latin1':
         return string.length;
@@ -305,7 +189,7 @@ export class Buffer extends Uint8Array {
       case 'utf16le':
         return string.length * 2;
       default:
-        return textEncoder.encode(string).byteLength;
+        return new TextEncoder().encode(string).byteLength;
     }
   }
